Split the mixed pilot test files along the window line
Three pilot test files hold both classes that drive a real top-level
window and classes that only build widgets, so the earlier move could
not take them whole. `pytest-fast` still pays for the window-driving
half of each, which is the cost the split directory exists to avoid.

Move only the `NO_LIVE_WINDOW`-guarded classes into `tests/gui/`:
`R2DWidgetSelectToolTC`, `R2DWidgetThemeTC`, and `PainterToolboxTC` to
`tests/gui/test_2d.py`, `DrawToolTC` to `tests/gui/test_draw_tool.py`,
and `PainterDesignCanvasTC` to `tests/gui/test_painter_design.py`. The
test bodies, their skip conditions, and their comments are unchanged;
only the file they live in moves.

Each new file carries its own imports and its own `NO_LIVE_WINDOW`, and
duplicates the module-level helpers it needs: `_load_rgba` in
`tests/gui/test_2d.py` and `_drawn_pixels` in
`tests/gui/test_draw_tool.py`, each still used by a class that stays
behind. Importing them across the two directories would couple the fast
and the GUI lane, which is what keeping them apart is for. `_send_mouse`
and `_click` had no remaining user and moved outright, as did the
imports and the `NO_LIVE_WINDOW` definitions the departing classes were
the only readers of.

Related to #1222
